requirements.txt
requests>=2.31.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.5.0
flake8>=6.1.0
bandit>=1.7.5
python-dotenv>=1.0.0

# Agno framework for multi-agent orchestration
agno>=2.0.0

# FastAPI for AgentOS runtime
fastapi>=0.104.0
uvicorn>=0.24.0

# Database support
sqlalchemy>=2.0.0

# Async support
aiohttp>=3.9.0
asyncio-compat